Fix: Obstacle Color Data

+++ Assets/Scripts/Grid/Logic/GroupChecker.cs
@@ -56,7 +56,7 @@ namespace GJG.GridSystem
                     foreach (int2 removeIndex in _toRemove)
                     {
                         _gridCheckFlag.Add(removeIndex);
-                        _painter.Paint(_gameGrid.GetItem(removeIndex), _gameGrid.GetNode(removeIndex).ItemBase.ColorType, currentItemType);
+                        _painter.Paint(_gameGrid.GetItem(removeIndex), _gameGrid.GetNode(removeIndex).ItemBase.ColorType, (int)currentItemType);
                     }
                 }
             }
@@ -94,7 +94,7 @@ namespace GJG.GridSystem
 
             foreach (int2 removeIndex in _toRemove)
             {
-                _painter.Paint(_gameGrid.GetItem(removeIndex), _gameGrid.GetNode(removeIndex).ItemBase.ColorType, GetItemType(_toRemove.Count));
+                _painter.Paint(_gameGrid.GetItem(removeIndex), _gameGrid.GetNode(removeIndex).ItemBase.ColorType, (int)GetItemType(_toRemove.Count));
             }
         }
